+++ CHANGELOG.md
@@ -7,5 +7,6 @@
   `RSpecFileName` and enabled them all by default ([@geniou][])
 * Add `RSpecDescription` cop to prevent to use of should in spec
   description. ([@geniou][])
+* `SpecFileName` cop for non-class specs ([@geniou][])
 
 [@geniou]: https://github.com/geniou